+++ scripts/dw/index.ts
@@ -1,10 +1,13 @@
 import { fatal } from "./helpers/shell.ts";
 import { cmdDefault } from "./commands/default.ts";
+import { cmdSetup } from "./commands/setup.ts";
+import { cmdRun } from "./commands/run.ts";
 import { cmdTeardown } from "./commands/teardown.ts";
 import { cmdList } from "./commands/list.ts";
 import { cmdReset } from "./commands/reset.ts";
 import { cmdLogs } from "./commands/logs.ts";
 import { cmdAttach } from "./commands/attach.ts";
+import { cmdIdentify } from "./commands/identify.ts";
 
 async function main(): Promise<void> {
 	const sub = process.argv[2];
@@ -13,6 +16,12 @@ async function main(): Promise<void> {
 		return;
 	}
 	switch (sub) {
+		case "setup":
+			await cmdSetup();
+			break;
+		case "run":
+			await cmdRun();
+			break;
 		case "teardown":
 			await cmdTeardown({ all: process.argv.includes("--all") });
 			break;
